Working on prefixexp
Justin Tomlinson
Oct 23, 2019

Lots of things are broken right now, but I have var working and
prefixexp is half working. Next is functioncall, and then I will work
on the easier stuff. I'm going to be spending a lot of time on this
today, so prepare for lots of changes.